attempt to use @babel/register with esm
Attempting to use @babel/register with esm led to esm's mainField
hook not working any more inside of a forked child process. I gave
up trying to get the local demo to work because @babel/register is
really hard to get working if it isn't run from the root now, and
it's not as important to get the demo working with Node 6.